sync local workspace

+++ pid_control_pkg/include/pid_control_pkg/pid_controller.hpp
@@ -7,8 +7,8 @@
 
 #include <rclcpp/rclcpp.hpp>
 #include <std_msgs/msg/bool.hpp>
+#include <std_msgs/msg/float32.hpp>
 #include <std_msgs/msg/float32_multi_array.hpp>
-#include <std_msgs/msg/int16.hpp>
 #include <std_msgs/msg/int32_multi_array.hpp>
 #include <tf2/LinearMath/Matrix3x3.h>
 #include <tf2/LinearMath/Quaternion.h>
@@ -64,7 +64,7 @@ public:
 
 private:
   void targetPositionCallback(const std_msgs::msg::Float32MultiArray::SharedPtr msg);
-  void heightCallback(const std_msgs::msg::Int16::SharedPtr msg);
+  void heightCallback(const std_msgs::msg::Float32::SharedPtr msg);
   void visualTakeoverCallback(const std_msgs::msg::Bool::SharedPtr msg);
   void fineDataCallback(const std_msgs::msg::Int32MultiArray::SharedPtr msg);
   void controlTimerCallback();
@@ -80,7 +80,7 @@ private:
   inline double radToDeg(double rad) const { return rad * 180.0 / M_PI; }
 
   rclcpp::Subscription<std_msgs::msg::Float32MultiArray>::SharedPtr target_position_sub_;
-  rclcpp::Subscription<std_msgs::msg::Int16>::SharedPtr height_sub_;
+  rclcpp::Subscription<std_msgs::msg::Float32>::SharedPtr height_sub_;
   rclcpp::Subscription<std_msgs::msg::Bool>::SharedPtr visual_takeover_sub_;
   rclcpp::Subscription<std_msgs::msg::Int32MultiArray>::SharedPtr fine_data_sub_;
   rclcpp::Publisher<std_msgs::msg::Float32MultiArray>::SharedPtr target_velocity_pub_;
